create JpaWithOptionsTestSupport, move support methods from JpaPagingTest
